[#986] refactoring adsense
