+++ app/assets/stylesheets/react.sass
@@ -1,19 +1,3 @@
 .react
   i.fa
     margin-right: 3.5px
-
-.google-adsense
-  p
-    margin: 0
-.rectangle-adsense
-  text-align: center
-  p
-    margin: 0
-
-.adslot_1
-  width: 270px
-  height: 250px
-@media (min-width:340px)
-  .adslot_1
-    width: 336px
-    height: 280px
