+++ app/views/layouts/_footer.html.slim
@@ -1,7 +1,7 @@
 footer.uk-container.uk-container-center.uk-margin.bottom
   .footer
     - if render_ads? && !Rails.env.test?
-      = rectangle_adsense
+      = adsense
     .relative
       .uk-panel.panel-default
         span.left-position
