+++ app/views/logs/sheet.html.slim
@@ -1,7 +1,7 @@
 h2 = fa_icon 'calendar-check-o', text: @title
 hr
 - unless user_signed_in? && current_user.special?
-  = rectangle_adsense
+  = adsense
   hr
 .uk-overflow-container
   table.uk-table.uk-table-hover.uk-table-striped.datatable
