+++ app/views/logs/show.html.slim
@@ -14,7 +14,7 @@ p
 = react_component 'TwitterSharedButton', text: "更新履歴(#{@logs.first.created_date})"
 hr
 - unless user_signed_in? && current_user.special?
-  = rectangle_adsense
+  = adsense
   hr
 
 = react_component 'LogCalendar', iidxid: params[:id]
