+++ app/views/sheets/power.html.slim
@@ -2,7 +2,7 @@
 h3.right サンプルデータはiidx.meより
 hr
 - unless user_signed_in? && current_user.special?
-  = rectangle_adsense
+  = adsense
   hr
 .uk-overflow-container
   table.uk-table.uk-table-hover.uk-table-striped.datatable
