+++ package.json
@@ -45,6 +45,7 @@
     "query-string": "^6.2.0",
     "rails-erb-loader": "^5.5.0",
     "react": "^16.6.3",
+    "react-adsense": "^0.0.6",
     "react-dom": "^16.6.3",
     "react-helmet": "^5.2.0",
     "react-redux": "^5.1.1",
